+++ R/predict.rma.r
@@ -786,6 +786,8 @@ predict.rma <- function(object, newmods, intercept, tau2.levels, gamma2.levels, 
       if (!inherits(vcovpred, "sparseMatrix"))
          class(vcovpred) <- c("vcovmat", class(vcovpred))
       out$vcov <- vcovpred
+      #attr(out$pred, ".rmspace") <- TRUE
+      #attr(out$vcov, ".rmspace") <- TRUE
    }
 
    return(out)