+++ tests/testthat/test_analysis_example_raudenbush2009.r
@@ -92,7 +92,7 @@ test_that("results are correct for the random-effects model (Knapp & Hartung met
 
    res.knha <- list()
 
-   expect_warning(res.knha$EE <- rma(yi, vi, data=dat, digits=3, method="EE", test="knha"))
+   res.knha$EE   <- rma(yi, vi, data=dat, digits=3, method="EE", test="knha")
    res.knha$ML   <- rma(yi, vi, data=dat, digits=3, method="ML", test="knha")
    res.knha$REML <- rma(yi, vi, data=dat, digits=3, method="REML", test="knha")
    res.knha$DL   <- rma(yi, vi, data=dat, digits=3, method="DL", test="knha")